working change
- working ShowResultSGPA.kt
- refactored CollegeChoose model name from an leftover refactor
