+++ app/src/main/java/com/xplorex/cgpabook/utils/Extensions.kt
@@ -1,6 +1,9 @@
 package com.xplorex.cgpabook.utils
 
+import android.content.ComponentName
 import android.content.Context
+import android.content.Intent
+import android.content.pm.ResolveInfo
 import android.net.ConnectivityManager
 import android.net.Uri
 import android.view.LayoutInflater
@@ -16,10 +19,10 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.SavedStateViewModelFactory
 import androidx.lifecycle.ViewModelProviders
 import com.android.volley.VolleyError
+import com.google.android.material.navigation.NavigationView
 import com.xplorex.cgpabook.R
 import com.xplorex.cgpabook.ui.SharedViewModel
 import com.xplorex.cgpabook.ui.profile.ProfileEditFragment
-import com.google.android.material.navigation.NavigationView
 import org.json.JSONObject
 import java.io.*
 
@@ -272,3 +275,50 @@ fun Fragment.goToUpdateProfile() {
         .addToBackStack(getString(R.string.menu_update_profile)).commit()
 }
 
+fun openAppRating(context: Context) {
+    // you can also use BuildConfig.APPLICATION_ID
+    val appId: String = context.packageName
+    val rateIntent = Intent(
+        Intent.ACTION_VIEW,
+        Uri.parse("market://details?id=$appId")
+    )
+    var marketFound = false
+
+    // find all applications able to handle our rateIntent
+    val otherApps: List<ResolveInfo> = context.packageManager
+        .queryIntentActivities(rateIntent, 0)
+    for (otherApp in otherApps) {
+        // look for Google Play application
+        if (otherApp.activityInfo.applicationInfo.packageName
+            == "com.android.vending"
+        ) {
+            val otherAppActivity = otherApp.activityInfo
+            val componentName = ComponentName(
+                otherAppActivity.applicationInfo.packageName,
+                otherAppActivity.name
+            )
+            // make sure it does NOT open in the stack of your activity
+            rateIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+            // task reparenting if needed
+            rateIntent.addFlags(Intent.FLAG_ACTIVITY_RESET_TASK_IF_NEEDED)
+            // if the Google Play was already open in a search result
+            //  this make sure it still go to the app page you requested
+            rateIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            // this make sure only the Google Play app is allowed to
+            // intercept the intent
+            rateIntent.component = componentName
+            context.startActivity(rateIntent)
+            marketFound = true
+            break
+        }
+    }
+
+    // if GP not present on device, open web browser
+    if (!marketFound) {
+        val webIntent = Intent(
+            Intent.ACTION_VIEW,
+            Uri.parse("https://play.google.com/store/apps/details?id=$appId")
+        )
+        context.startActivity(webIntent)
+    }
+}
